+++ CHANGELOG.md
@@ -48,6 +48,8 @@ This project adheres to [Semantic Versioning](http://semver.org/).
   `::nix::sys::reboot` now return `Result<Infallible>` instead of `Result<Void>` (#[1239](https://github.com/nix-rust/nix/pull/1239))
 - `sys::socket::sockaddr_storage_to_addr` is no longer `unsafe`.  So is
   `offset_of!`.
+- `sys::socket::sockaddr_storage_to_addr`, `offset_of!`, and `Errno::clear` are
+  no longer `unsafe`.
   (#[1244](https://github.com/nix-rust/nix/pull/1244))
 - Several `Inotify` methods now take `self` by value instead of by reference
   (#[1244](https://github.com/nix-rust/nix/pull/1244))